Remove direct-access of ObjectType children.
ObjectType was the last microstate type to eagerly compute all of the
keys of its children. This made it not only run counter to every other
microstate which are completely lazy, but it also meant that its
performance was abysmal.

For this reason, as well as others, direct-access to children was
removed in 0.14.0. If you still need to access the children of an
Object type directly, you can use the `entries` hash. But the
preferred method is through enumeration.

Note that more utilities will be added in subsequent versions of
microstates to make accessing children of the object type eaiser and
still more performant.

See https://github.com/microstates/microstates.js/pull/338 for details.